requirements.txt
pandas
numpy
scikit-learn
pymongo>=4.6
seaborn
streamlit
tensorflow==2.12.1
gensim
nltk
python-dotenv
certifi
nltk
gensim
joblib
